+++ src/components/BurgerConstructorElement/BurgerConstructorElement.jsx
@@ -46,7 +46,21 @@ const BurgerConstructorElement = (props) => {
   };
   
   BurgerConstructorElement.propTypes = {
-    data : PropTypes.array.isRequired
+    data: PropTypes.arrayOf(
+      PropTypes.shape({
+        _id: PropTypes.string.isRequired,
+        name: PropTypes.string.isRequired,
+        type: PropTypes.string.isRequired,
+        proteins: PropTypes.number,
+        fat: PropTypes.number,
+        carbohydrates: PropTypes.number,
+        calories: PropTypes.number,
+        price: PropTypes.number.isRequired,
+        image: PropTypes.string,
+        image_mobile: PropTypes.string,
+        image_large: PropTypes.string,
+      })
+    ).isRequired
   }
