+++ src/components/BurgerIngredients/Burgeringridients.jsx
@@ -63,7 +63,21 @@ const BurgerIngredients = (props) => {
 }
 
 BurgerIngredients.propTypes = {
-data: PropTypes.array.isRequired
+  data: PropTypes.arrayOf(
+    PropTypes.shape({
+      _id: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
+      type: PropTypes.string.isRequired,
+      proteins: PropTypes.number,
+      fat: PropTypes.number,
+      carbohydrates: PropTypes.number,
+      calories: PropTypes.number,
+      price: PropTypes.number.isRequired,
+      image: PropTypes.string,
+      image_mobile: PropTypes.string,
+      image_large: PropTypes.string,
+    })
+  ).isRequired
 }
