+++ src/components/ItemBlock/ItemBlock.jsx
@@ -38,6 +38,20 @@ const ItemBlock = (props) => {
 
 
   ItemBlock.propTypes = {
-    data: PropTypes.array
+    data: PropTypes.arrayOf(
+      PropTypes.shape({
+        _id: PropTypes.string.isRequired,
+        name: PropTypes.string.isRequired,
+        type: PropTypes.string.isRequired,
+        proteins: PropTypes.number,
+        fat: PropTypes.number,
+        carbohydrates: PropTypes.number,
+        calories: PropTypes.number,
+        price: PropTypes.number.isRequired,
+        image: PropTypes.string,
+        image_mobile: PropTypes.string,
+        image_large: PropTypes.string,
+      })
+    ).isRequired
   }
   export default ItemBlock;
