+++ src/components/ItemBlock/ItemBlock.module.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-direction: column;
     gap: 10px;
-    height: 570px;
+    height: 440px;
     margin: 8px 0px;
     overflow: auto;
 }
